requirements.txt
setuptools
Django==3.1
djangorestframework==3.11.1
django-rest-auth==0.9.5
django-cors-headers==3.5.0
psycopg2-binary
pandas
scikit-learn
joblib
gunicorn